requirements.txt
pillow>=9.0.0
tensorflow>=2.7.0
huggingface_hub>=0.0.19
fastapi>=0.100.0
uvicorn>=0.22.0
git+https://github.com/KichangKim/DeepDanbooru@v3-20200915-sgd-e30#egg=deepdanbooru